cli: move dev commmands to server
